Update interact button images
Because the interact functionality depends on knowing what the "Enter"
and "Change" buttons look like, it requires that the images match what
appears in game (at all 3 quality levels) exactly. At some point, the
game must have updated the images, as they no longer matched what I had
since the interact feature's initial release. An update of the images
solved that problem.

The code change is merely to take into account the image size (which
is ever so slightly larger than the buttons' sizes), and click in the
middle of the buttons rather than on the edges. These are effectively
magic numbers, and should be explained in variable names in the future.

Fixes #25